refactor(imports): use src alias across app and fix Routes/Modal
