+++ src/features/app/__tests__/App.spec.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import App from 'src/features/app';
+import App from 'src/features/app/App';
 import store from 'src/store/configureStore';
 
 describe('App', () => {
