+++ src/features/common/components/DatePicker/index.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import moment from 'moment';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
-import './DatePicker.scss';
+import 'src/features/common/components/DatePicker/DatePicker.scss';
 
 /**
  * @exports
