+++ src/features/not-found/__tests__/NotFound.spec.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import NotFound from 'src/features/not-found';
+import NotFound from 'src/features/not-found/NotFound';
 
 describe('NotFound', () => {
   it('should render NotFound component correctly', () => {
